django_adlamandnko/pyproject.toml
[tool.poetry]
name = "adlamandnko"
version = "0.1.0"
description = "Website for Adlam and N'ko languages annotation"
authors = ["Boubacar, <boubasow.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-bootstrap-themes = "^3.3.6"
tree = "^0.2.4"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
keyboard = "^0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
